third_party/blink/web_tests/external/wpt/html/cross-origin-opener-policy/coep-navigate-popup.https_0-1-expected.txt

This is a testharness.js-based test.
PASS Popup navigating to same-origin with coop/coep
FAIL Popup navigating to same-site with coop/coep assert_equals: Openee browsing context closed expected true but got false
FAIL Popup navigating to same-origin with no coop/coep assert_equals: Openee browsing context closed expected true but got false
FAIL Popup navigating to same-site with no coop/coep assert_equals: Openee browsing context closed expected true but got false
Harness: the test ran to completion.
